requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
python-telegram-bot==12.3.0